pyproject.toml
[tool.poetry]
name = "uofg-itech-project"
version = "0.1.0"
description = ""
authors = ["Pedro Belfort <belfortdev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
django-webpack-loader = "^0.6.0"
djangorestframework = "^3.11.0"
Pillow = "^7.0.0"
drf-extensions = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
